# sacdia-api-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/auth/mfa/enroll": {
      "post": {
        "description": "Genera un QR code y secret para configurar en tu app de autenticación",
        "operationId": "MfaController_enrollMfa_v1",
        "parameters": [
          {
            "name": "x-refresh-token",
            "in": "header",
            "description": "Refresh token opcional para ligar sesión MFA en entornos que lo requieran.",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "QR code y secret generados",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "factorId": {
                      "type": "string"
                    },
                    "qrCode": {
                      "type": "string",
                      "description": "Base64 del QR code"
                    },
                    "secret": {
                      "type": "string",
                      "description": "Secret para configurar manualmente"
                    },
                    "uri": {
                      "type": "string",
                      "description": "URI para apps de autenticación"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Iniciar enrolamiento de 2FA",
        "tags": [
          "auth"
        ]
      }
    },
    "/v1/auth/mfa/verify": {
      "post": {
        "description": "Verifica el código TOTP y activa 2FA para la cuenta",
        "operationId": "MfaController_verifyMfa_v1",
        "parameters": [
          {
            "name": "x-refresh-token",
            "in": "header",
            "description": "Refresh token opcional para ligar sesión MFA en entornos que lo requieran.",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyMfaDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "2FA activado exitosamente"
          },
          "401": {
            "description": "Código inválido"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Verificar y activar 2FA",
        "tags": [
          "auth"
        ]
      }
    },
...
    "/v1/auth/mfa/status": {
      "get": {
        "description": "Indica si el usuario tiene 2FA habilitado y su nivel de autenticación",
        "operationId": "MfaController_getMfaStatus_v1",
        "parameters": [
          {
            "name": "x-refresh-token",
            "in": "header",
            "description": "Refresh token opcional para ligar sesión MFA en entornos que lo requieran.",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "mfaEnabled": {
                      "type": "boolean"
                    },
                    "currentLevel": {
                      "type": "string",
                      "description": "aal1 = password, aal2 = password + MFA"
                    },
                    "factors": {
                      "type": "array"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Verificar estado de 2FA",
        "tags": [
          "auth"
        ]
      }
    },
...
  },
  "info": {
    "title": "SACDIA API",
    "description": "## Sistema de Administración de Clubes de Conquistadores y Aventureros\n\n### Módulos Disponibles\n- **Auth**: Autenticación con Supabase + JWT\n- **Users**: Gestión de perfiles de usuario\n- **Catalogs**: Catálogos de referencia\n- **Clubs**: Gestión de clubes e instancias\n- **Classes**: Clases progresivas\n- **Honors**: Especialidades\n- **Activities**: Actividades de club\n- **Finances**: Control financiero",
    "version": "2.2.0",
    "contact": {
      "name": "SACDIA Team",
      "url": "https://sacdia.app",
      "email": "dev@sacdia.app"
    }
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "VerifyMfaDto": {
        "type": "object",
        "properties": {
          "factorId": {
            "type": "string",
            "example": "abc123-factor-id",
            "description": "ID del factor MFA a verificar"
          },
          "code": {
            "type": "string",
            "example": "123456",
            "description": "Código TOTP de 6 dígitos de tu app de autenticación"
          }
        },
        "required": [
          "factorId",
          "code"
        ]
      },
...
    }
  }
}